declared controller fs for products router; moved fs from routes to controller fs
